Fix TS compiler errors in components after update

+++ src/app/pages/patients/glucose-levels/observation-form/observation-form.component.ts
@@ -16,10 +16,10 @@ import * as resourceUtils from "../../../../@core/services/utils/resource-utils"
   styleUrls: ['./observation-form.component.scss']
 })
 export class ObservationFormComponent extends FormComponent implements OnInit {
-  @Input() observation: Observation;
+  @Input() observation: Observation | undefined;
   @Input() isUpdate: boolean = false;
   timesOfDay = TimeOfDay;
-  observationForm: FormGroup;
+  observationForm: FormGroup | undefined;
   saved: boolean = false;
   localeTime = 'dd/MM/yyyy HH:mm'
   showDeleteMessage: boolean = false;
@@ -46,19 +46,24 @@ export class ObservationFormComponent extends FormComponent implements OnInit {
   }
 
   get valueControl(): FormControl {
-    return this.observationForm.get('value') as FormControl;
+    return this.observationForm?.get('value') as FormControl;
   }
 
   get dateControl(): FormControl {
-    return this.observationForm.get('date') as FormControl;
+    return this.observationForm?.get('date') as FormControl;
   }
 
   get timingControl(): FormControl {
-    return this.observationForm.get('timing') as FormControl;
+    return this.observationForm?.get('timing') as FormControl;
   }
 
   submitForm(): void {
-    this.observation.valueQuantity.value = this.valueControl.value;
+    if (!this.observation) {
+      this.formError('Observation not defined');
+      return;
+    }
+
+    this.observation.valueQuantity = {value: this.valueControl.value};
     this.observation.issued = this.dateControl.value.toISOString();
     resourceUtils.setCodeExtension(this.observation, Extensions.RESOURCE_TIMING, this.timingControl.value);
 
@@ -75,6 +80,11 @@ export class ObservationFormComponent extends FormComponent implements OnInit {
   }
 
   deleteRecord(): void {
+    if (!this.observation?.id) {
+      this.formError('Observation ID not defined');
+      return;
+    }
+
     this.formStatus = FormStatus.loading;
     this.observationService.deleteObservation(this.observation.id)
       .subscribe(this.formSuccess, this.formError)
@@ -85,7 +95,7 @@ export class ObservationFormComponent extends FormComponent implements OnInit {
     this.saved = true;
   }
 
-  private formError = (error) => {
+  private formError = (error: any) => {
     this.formStatus = FormStatus.error;
     console.log(error);
   }
